verification/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
Pillow
numpy
